+++ src/app/api.js
@@ -39,7 +39,6 @@ const driesBase = {
   description: driesDescription,
   layers: [
     'cells',
-    'factors',
   ].map(makeLayerNameToConfig('dries')),
 };
 
@@ -343,9 +342,9 @@ const configs = {
         props: {
           description: driesDescription,
         },
-        x: 0, y: 0, w: 5, h: 2 },
+        x: 0, y: 0, w: 3, h: 2 },
       { component: 'status',
-        x: 0, y: 1, w: 5, h: 2 },
+        x: 0, y: 1, w: 3, h: 2 },
       { component: 'scatterplot',
         props: {
           mapping: 't-SNE',
@@ -354,7 +353,7 @@ const configs = {
             target: [0, 0, 0],
           },
         },
-        x: 0, y: 2, w: 5, h: 4 },
+        x: 0, y: 2, w: 3, h: 4 },
       { component: 'spatial',
         props: {
           cellRadius: 50,
@@ -363,7 +362,7 @@ const configs = {
             target: [3800, -900, 0],
           },
         },
-        x: 5, y: 0, w: 5, h: 4 },
+        x: 3, y: 0, w: 3, h: 4 },
       { component: 'scatterplot',
         props: {
           mapping: 'UMAP',
@@ -372,9 +371,10 @@ const configs = {
             target: [0, 0, 0],
           },
         },
-        x: 5, y: 4, w: 5, h: 4 },
-      { component: 'factors',
-        x: 10, y: 0, w: 2, h: 8 },
+        x: 3, y: 4, w: 3, h: 4 },
+      { component: 'group-selector',
+        props: {},
+        x: 6, y: 0, w: 6, h: 8 },
     ],
   },
   'wang-2019': {