+++ src/app/componentRegistry.js
@@ -9,6 +9,7 @@ import HoverableScatterplotSubscriber from '../components/scatterplot/HoverableS
 import HoverableSpatialSubscriber from '../components/spatial/HoverableSpatialSubscriber';
 import HoverableHeatmapSubscriber from '../components/heatmap/HoverableHeatmapSubscriber';
 import ChannelsSubscriber from '../components/channels/ChannelsSubscriber';
+import GroupSelectorSubscriber from '../components/group-selector/GroupSelectorSubscriber';
 
 class Description extends React.Component {
   componentDidMount() {
@@ -36,6 +37,7 @@ const registry = {
   spatial: HoverableSpatialSubscriber,
   heatmap: HoverableHeatmapSubscriber,
   channels: ChannelsSubscriber,
+  'group-selector': GroupSelectorSubscriber,
 };
 
 export default function getComponent(name) {